introduce Flix.addFpkg
